Move converters registry implementation to separate package
